Improve readability of welcome screen on small screens.
The features are hard to read on smaller device sizes (e.g. iPhone SE), since the scroll area is very small.

To alleviate this, the title is incorporated in the scroll view and the spacer above the button explanation as well. This leaves a larger scrollable area which results in a more pleasant reading experience.