Fix bugs due to change in Codable representation of CompendiumItemKey
The modularisation effort made CompendiumItemKey no longer RawRepresentable,
which caused its Codable representation to change from a String to a Dictionary.
This was fixed by using @Migrated in CompendiumItemReference (the only
context in which CompendiumItemKey is persisted).

This commit refactors how the CompendiumEntry's db key and CompendiumItemKey
relate to each other. They used to be very intertwined.